Separate Filteriffic interface into separate files
