+++ tgui/packages/tgui/interfaces/Filteriffic/filterDataEntries/FilterBlendmodeEntry.tsx
@@ -1,0 +1,35 @@
+/**
+ * @file
+ * @copyright 2020
+ * @author actioninja (https://github.com/actioninja)
+ * @license MIT
+ */
+
+import { map } from 'common/collections';
+import { Button } from 'tgui-core/components';
+
+import { useBackend } from '../../../backend';
+import type { FilterifficData } from '../type';
+
+export const FilterBlendmodeEntry = (props) => {
+  const { name, value, filterName, filterType } = props;
+  const { act, data } = useBackend<FilterifficData>();
+
+  const filterInfo = data.filter_info;
+  const flags = filterInfo[filterType]['blend_mode'];
+  return map(flags, (flagField, flagName) => (
+    <Button.Checkbox
+      checked={value === flagField}
+      onClick={() =>
+        act('modify_filter_value', {
+          name: filterName,
+          new_data: {
+            [name]: flagField,
+          },
+        })
+      }
+    >
+      {flagName}
+    </Button.Checkbox>
+  ));
+};
