+++ tgui/packages/tgui/interfaces/Filteriffic/filterDataEntries/FilterIconEntry.tsx
@@ -1,0 +1,31 @@
+/**
+ * @file
+ * @copyright 2020
+ * @author actioninja (https://github.com/actioninja)
+ * @license MIT
+ */
+
+import { Box, Button } from 'tgui-core/components';
+
+import { useBackend } from '../../../backend';
+import type { FilterifficData } from '../type';
+
+export const FilterIconEntry = (props) => {
+  const { value, filterName } = props;
+  const { act } = useBackend<FilterifficData>();
+  return (
+    <>
+      <Button
+        icon="pencil-alt"
+        onClick={() =>
+          act('modify_icon_value', {
+            name: filterName,
+          })
+        }
+      />
+      <Box inline ml={1}>
+        {value}
+      </Box>
+    </>
+  );
+};
