+++ tgui/packages/tgui/interfaces/Filteriffic/filterDataEntries/index.ts
@@ -1,0 +1,16 @@
+/**
+ * @file
+ * @copyright 2025
+ * @author Mordent (https://github.com/mordent-goonstation)
+ * @license ISC
+ */
+
+export { FilterBlendmodeEntry } from './FilterBlendmodeEntry';
+export { FilterColorEntry } from './FilterColorEntry';
+export { FilterFlagsEntry } from './FilterFlagsEntry';
+export { FilterFloatEntry } from './FilterFloatEntry';
+export { FilterIconEntry } from './FilterIconEntry';
+export { FilterIntegerEntry } from './FilterIntegerEntry';
+export { FilterSpaceEntry } from './FilterSpaceEntry';
+export { FilterTextEntry } from './FilterTextEntry';
+export { FilterTransformEntry } from './FilterTransformEntry';
