+++ tgui/packages/tgui/interfaces/Filteriffic/type.ts
@@ -1,0 +1,12 @@
+/**
+ * @file
+ * @copyright 2025
+ * @author Mordent (https://github.com/mordent-goonstation)
+ * @license MIT
+ */
+
+export interface FilterifficData {
+  filter_info;
+  target_name;
+  target_filter_data;
+}
